\name{RcppCorels-package}
\alias{RcppCorels-package}
\alias{RcppCorels}
\docType{package}
\title{\packageTitle{RcppCorels}}
\description{\packageDescription{RcppCorels}}
\section{Package Content}{\packageIndices{RcppCorels}}
\author{\packageAuthor{RcppCorels}}
\section{Maintainer}{\packageMaintainer{RcppCorels}}
\keyword{package}
